+++ src/characters/ursula.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { PolicyMessageKit } from '../crypto/kits';
+import { PolicyMessageKit } from '../kits/message';
 import { Arrangement } from '../policies/policy';
 import { ChecksumAddress } from '../types';
