+++ src/kits/message.ts
@@ -1,23 +1,28 @@
-import { Capsule, CapsuleWithFrags, PublicKey, Signer } from 'umbral-pre';
+import {
+  Capsule,
+  CapsuleWithFrags,
+  encrypt,
+  PublicKey,
+  Signer,
+} from 'umbral-pre';
 
 import { Enrico } from '../characters/enrico';
-import { PrePublishedTreasureMap, Revocation } from '../policies/collections';
+import { Revocation } from '../policies/collections';
 import { ChecksumAddress } from '../types';
 
-export class RevocationKit {
-  public revocations: Record<ChecksumAddress, Revocation>;
-
-  constructor(treasureMap: PrePublishedTreasureMap, signer: Signer) {
-    this.revocations = {};
-    Object.entries(treasureMap.destinations).forEach(
-      ([nodeId, arrangementId]) => {
-        this.revocations[nodeId] = new Revocation(arrangementId, signer);
-      }
-    );
-  }
+export interface MessageKit {
+  capsule: Capsule;
+  ciphertext: Buffer;
+  signature: Buffer;
+  senderVerifyingKey: PublicKey;
+  recipientEncryptingKey: PublicKey;
 }
 
-export class PolicyMessageKit {
+export type ReencryptedMessageKit = Omit<MessageKit, 'capsule'> & {
+  capsule: CapsuleWithFrags;
+};
+
+export class PolicyMessageKit implements MessageKit {
   public readonly capsule: Capsule;
   public readonly ciphertext: Buffer;
   public readonly signature: Buffer;
@@ -29,7 +34,7 @@ export class PolicyMessageKit {
     ciphertext: Buffer,
     signature: Buffer,
     senderVerifyingKey: PublicKey,
-    recipientEncryptingKey: PublicKey
+    recipientEncryptingKey: PublicKey,
   ) {
     this.capsule = capsule;
     this.ciphertext = ciphertext;
@@ -39,7 +44,7 @@ export class PolicyMessageKit {
   }
 
   public toBytes(includeAlicePublicKey = true): Buffer {
-    const asBytes = [this.capsule.toBytes()];
+    const asBytes = [ this.capsule.toBytes() ];
     if (includeAlicePublicKey && !!this.senderVerifyingKey) {
       asBytes.push(this.senderVerifyingKey.toBytes());
     }
@@ -49,7 +54,7 @@ export class PolicyMessageKit {
 
   public ensureCorrectSender(
     enrico: Enrico,
-    recipientEncryptingKey: PublicKey
+    recipientEncryptingKey: PublicKey,
   ): void {
     if (
       enrico.recipientEncryptingKey !== this.recipientEncryptingKey &&
@@ -62,17 +67,4 @@ export class PolicyMessageKit {
     //   throw new Error('Sender verifying key does not match');
     // }
   }
-}
-
-// Just like a PolicyMessageKit, but with a different `capsule` type
-// TODO: Cant use re-use existing type because of missing `toBytes` implementation
-// export type ReencryptedMessageKit = Omit<PolicyMessageKit, 'capsule'> & {
-//   capsule: CapsuleWithFrags;
-// };
-export interface ReencryptedMessageKit {
-  capsule: CapsuleWithFrags;
-  ciphertext: Buffer;
-  signature: Buffer;
-  senderVerifyingKey: PublicKey;
-  recipientEncryptingKey: PublicKey;
 }
